Split streaming helper tests
Move streaming attribute helper tests out of helpers.rs and assert the forwarded unknown-entity normalization error explicitly.
